IGNITE-24261 Moved string_utils.h to common
