+++ modules/platforms/cpp/ignite/common/detail/string_utils.h
