+++ modules/platforms/cpp/ignite/odbc/config/connection_info.cpp
@@ -16,7 +16,7 @@
  */
 
 #include "ignite/odbc/config/connection_info.h"
-#include "ignite/odbc/string_utils.h"
+#include "ignite/common/detail/string_utils.h"
 #include "ignite/odbc/system/odbc_constants.h"
 #include "ignite/odbc/utility.h"
